+++ RELEASE_NOTES.md
@@ -39,6 +39,7 @@ ns-3 has switched to the C++23 standard by default.
 ### Bugs fixed
 
 - (internet) #1251 - Added check for longest prefix match in GlobalRouting
+- (wifi) Block transmission on other EMLSR links as soon as it is detected that the main PHY is receiving an A-MPDU, to prevent that the EMLSR client starts an UL TXOP before the end of the A-MPDU
 
 ## Release 3.45